Added new mapping system based on annotations, added database credentials saving.

+++ Backend/src/main/java/com/server/Server.java
@@ -26,7 +26,7 @@ public interface Server {
 
     ChannelFuture getChannel();
 
-    short getPort();
+    int getPort();
 
     Environment getEnvironment();
 }
